requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
alembic==1.13.2
annotated-types==0.7.0
anyio==3.7.1
apache-airflow==2.10.2
apache-airflow-providers-common-compat==1.2.0
apache-airflow-providers-common-io==1.4.0
apache-airflow-providers-common-sql==1.4.0
apache-airflow-providers-fab==1.3.0
apache-airflow-providers-ftp==3.3.1
apache-airflow-providers-http==4.3.0
apache-airflow-providers-imap==3.1.1
apache-airflow-providers-smtp==1.8.0
apache-airflow-providers-sqlite==3.3.2
apache-airflow-providers-ssh==3.13.1
apache-airflow-providers-yandex==3.12.0
apispec==6.6.1
appnope==0.1.4
argcomplete==1.12.3
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.8.1
asttokens==2.4.1
async-lru==2.0.4
async-timeout==4.0.3
attrs==24.2.0
babel==2.16.0
backcall==0.2.0
bcrypt==4.2.0
beautifulsoup4==4.12.3
bleach==6.1.0
blinker==1.8.2
boto3==1.35.24
botocore==1.35.24
cachelib==0.9.0
cachetools==5.5.0
cattrs==1.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
clickclick==20.10.2
cloudpickle==3.1.0
colorama==0.4.6
colorlog==6.8.2
comm==0.2.2
ConfigUpdater==3.2
connexion==2.14.2
contourpy==1.1.1
cron-descriptor==1.4.5
croniter==3.0.3
cryptography==42.0.8
cycler==0.12.1
databricks-sdk==0.35.0
debugpy==1.8.7
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.14
dill==0.3.8
dnspython==2.6.1
docker==7.1.0
docutils==0.16
email-validator==2.2.0
exceptiongroup==1.2.2
executing==2.1.0
fastapi==0.115.4
fastjsonschema==2.20.0
findspark==2.0.1
Flask==2.2.5
Flask-AppBuilder==4.5.0
Flask-Babel==2.0.0
Flask-Caching==2.3.0
Flask-JWT-Extended==4.6.0
Flask-Limiter==3.8.0
Flask-Login==0.6.3
Flask-OpenID==1.3.1
flask-session==0.5.0
Flask-SQLAlchemy==2.5.1
flask-wtf==1.2.1
fonttools==4.54.1
fqdn==1.5.1
frozenlist==1.4.1
fsspec==2024.9.0
gitdb==4.0.11
GitPython==3.1.43
google-auth==2.35.0
google-re2==1.1.20240702
googleapis-common-protos==1.65.0
graphene==3.4
graphql-core==3.2.5
graphql-relay==3.2.0
graphviz==0.20.3
grpcio==1.66.1
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
idna==3.10
importlib-metadata==8.4.0
importlib-resources==5.13.0
inflection==0.5.1
iniconfig==2.0.0
ipykernel==6.29.5
ipython==8.12.3
ipywidgets==8.1.5
iso8601==2.1.0
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.1
jinja2==3.1.4
jmespath==1.0.1
joblib==1.4.2
json5==0.9.25
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter==1.1.1
jupyter-client==8.6.3
jupyter-console==6.6.3
jupyter-core==5.7.2
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter-server==2.14.2
jupyter-server-terminals==0.5.3
jupyterlab==4.2.5
jupyterlab-pygments==0.3.0
jupyterlab-server==2.27.3
jupyterlab-widgets==3.0.13
kafka-python==2.0.2
kiwisolver==1.4.7
lazy-object-proxy==1.10.0
limits==3.13.0
linkify-it-py==2.0.3
lockfile==0.12.2
Mako==1.3.5
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.22.0
marshmallow-enum==1.5.1
marshmallow-oneofschema==3.1.1
marshmallow-sqlalchemy==0.26.1
matplotlib==3.7.5
matplotlib-inline==0.1.7
mdit-py-plugins==0.4.2
mdurl==0.1.2
methodtools==0.4.7
mistune==3.0.2
mlflow==2.17.0
mlflow-skinny==2.17.0
multidict==6.1.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
notebook==7.2.2
notebook-shim==0.2.4
numpy==1.24.4
openapi-schema-validator==0.2.3
openapi-spec-validator==0.4.0
opentelemetry-api==1.27.0
opentelemetry-exporter-otlp==1.27.0
opentelemetry-exporter-otlp-proto-common==1.27.0
opentelemetry-exporter-otlp-proto-grpc==1.27.0
opentelemetry-exporter-otlp-proto-http==1.27.0
opentelemetry-proto==1.27.0
opentelemetry-sdk==1.27.0
opentelemetry-semantic-conventions==0.48b0
ordered-set==4.1.0
overrides==7.7.0
packaging==24.1
pandas==2.0.3
pandocfilters==1.5.1
paramiko==3.5.0
parso==0.8.4
pathspec==0.12.1
pendulum==2.1.2
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.4.0
pkgutil-resolve-name==1.3.10
platformdirs==4.3.6
pluggy==1.5.0
prison==0.2.1
prometheus-client==0.21.0
prompt-toolkit==3.0.48
protobuf==4.25.5
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.3
py4j==0.10.9
pyarrow==17.0.0
pyasn1==0.6.1
pyasn1-modules==0.4.1
pycparser==2.22
pydantic==2.9.2
pydantic-core==2.23.4
pygments==2.18.0
PyJWT==2.9.0
PyNaCl==1.5.0
pyparsing==3.1.4
pyrsistent==0.20.0
pyspark==3.0.2
pyspark-stubs==3.0.0.post3
pytest==8.3.3
python-daemon==3.0.1
python-dateutil==2.9.0.post0
python-json-logger==2.0.7
python-nvd3==0.15.0
python-slugify==8.0.4
python3-openid==3.2.0
pytz==2024.2
pytzdata==2020.1
PyYAML==6.0.2
pyzmq==26.2.0
referencing==0.35.1
requests==2.32.3
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986==1.5.0
rfc3986-validator==0.1.1
rich==13.8.1
rich-argparse==1.5.2
rpds-py==0.20.0
rsa==4.9
s3transfer==0.10.2
scikit-learn==1.3.2
scipy==1.10.1
Send2Trash==1.8.3
setproctitle==1.3.3
sh==2.0.7
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==1.4.54
SQLAlchemy-JSONField==1.0.2
SQLAlchemy-Utils==0.41.2
sqlparse==0.5.1
sshtunnel==0.4.0
stack-data==0.6.3
starlette==0.41.2
swagger-ui-bundle==0.0.9
tabulate==0.8.10
tenacity==9.0.0
termcolor==2.4.0
terminado==0.18.1
text-unidecode==1.3
threadpoolctl==3.5.0
tinycss2==1.3.0
tomli==2.0.2
tornado==6.4.1
traitlets==5.14.3
types-python-dateutil==2.9.0.20241003
typing-extensions==4.12.2
tzdata==2024.2
uc-micro-py==1.0.3
unicodecsv==0.14.1
universal-pathlib==0.2.5
uri-template==1.3.0
urllib3==1.26.20
wcwidth==0.2.13
webcolors==24.8.0
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==2.2.3
widgetsnbextension==4.0.13
wirerope==0.4.7
wrapt==1.16.0
WTForms==2.3.3
yandex-query-client==0.1.4
yandexcloud==0.319.0
yarl==1.11.1
zipp==3.20.2